client/requirements.txt
streamlit>=1.31.0
streamlit-chat>=0.1.1
httpx>=0.26.0
python-dotenv>=1.0.0